Restructured tests to match new library structure
