+++ pyeo/tests/context.py
@@ -1,3 +1,2 @@
 import sys, os
 sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..')))
-from pyeo import core as pyeo
