--- pyeo/tests/example_project_tests.py
@@ -1,4 +1,0 @@
-import os, sys
-import shutil
-sys.path.insert(0, os.path.abspath(os.path.join(__file__, '..', '..', '..')))
-import pyeo.core as pyeo
